# Fix escape chars in output when `linearize: False`